requirements.txt
asgiref==3.10.0
bleach==6.3.0
Django==5.2.7
django-debug-toolbar==6.1.0
django-extensions==4.1
django-multi-captcha-admin==2.0.0
django-ranged-response==0.2.0
django-robots==6.1
django-simple-captcha==0.6.2
django-summernote==0.8.20.0
django-taggit==6.1.0
pillow==12.0.0
sqlparse==0.5.3
tzdata==2025.2
webencodings==0.5.1
